refactor pm quality score run detail router
